Combine GenerateStats and TfDataValidator

+++ tests/tfdv_test.py
@@ -18,17 +18,21 @@
 
 import os
 import tempfile
+from os.path import join as pjoin
 from unittest import TestCase
 
 from examples.examples_utils import get_taxi_data_dir
-from spotify_tensorflow.tfx.tfdv import GenerateStats
+from spotify_tensorflow.tfx.tfdv import TfDataValidator
 
 
-class TFDVTest(TestCase):
+class TfDataValidatorTest(TestCase):
     def setUp(self):
         taxi_data = get_taxi_data_dir()
-        self.input_data_dir = taxi_data
-        self.stats_file = os.path.join(taxi_data, "stats.pb")
+        self.data_location = taxi_data
+        self.schema = pjoin(taxi_data, "chicago_taxi_schema.pbtxt")
+        self.schema_snapshot_path = pjoin(taxi_data, "schema_snapshot.pb")
+        self.stats_file = pjoin(taxi_data, "stats.pb")
+        self.anomalies_path = pjoin(self.data_location, "anomalies.pb")
         tmp_dir = tempfile.mkdtemp()
         self.pipeline_args = [
             "--temp_location=%s" % tmp_dir,
@@ -36,9 +40,23 @@ class TFDVTest(TestCase):
             "--runner=DirectRunner"
         ]
 
-    def test_generate_stats(self):
-        GenerateStats(self.input_data_dir).run(self.pipeline_args)
+    def test_write_stats_and_schema(self):
+        TfDataValidator(self.schema, self.data_location).write_stats_and_schema(self.pipeline_args)
         self.assertTrue(os.path.exists(self.stats_file))
+        self.assertTrue(os.path.exists(self.schema_snapshot_path))
+
+    def test_validate_stats_against_schema(self):
+        validator = TfDataValidator(self.schema, self.data_location)
+        validator.write_stats(self.pipeline_args)
+        self.assertTrue(os.path.exists(self.stats_file))
+        has_no_anomalies = validator.validate_stats_against_schema()
+        self.assertFalse(has_no_anomalies)
+        self.assertTrue(os.path.exists(self.anomalies_path))
 
     def tearDown(self):
-        os.remove(self.stats_file)
+        if os.path.exists(self.stats_file):
+            os.remove(self.stats_file)
+        if os.path.exists(self.schema_snapshot_path):
+            os.remove(self.schema_snapshot_path)
+        if os.path.exists(self.anomalies_path):
+            os.remove(self.anomalies_path)
